requirements.txt
openai
pinecone
python-dotenv
fastapi
uvicorn
tiktoken
supabase
